documenting wallet currency relatd util functions in the walletCurrencies module
