+++ js/views/modals/orderDetail/summaryTab/DisputeStarted.js
@@ -1,6 +1,6 @@
 import _ from 'underscore';
 import moment from 'moment';
-import { getServerCurrency } from '../../../../data/cryptoCurrencies';
+import { getServerCurrency } from '../../../../data/walletCurrencies';
 import {
   events as orderEvents,
 } from '../../../../utils/order';
